OSIS-6110 : Remove/comment useless
